changed swaggerDocumentation.yaml to use Long for IDs of images

+++ backend/src/test/java/at/ac/tuwien/sepm/groupphase/backend/unittests/Article/ArticleValidatorTest.java
@@ -38,8 +38,8 @@ class ArticleValidatorTest {
         articleWithoutIdDto.setTitle("   ");
         articleWithoutIdDto.setSummary("   ");
         articleWithoutIdDto.setText("   ");
-        List<Integer> images = new ArrayList<>();
-        images.add(1);
+        List<Long> images = new ArrayList<>();
+        images.add(1L);
 
         articleWithoutIdDto.setImages(images);
